Implement several missing methods.
Based on the patch by Cosimo Cecchi.

This adds the stubs for proper net status notification,
implements get_title, get_location and workarounds the bug
in webkit that will make pages not load images if the protocol
is not specified in the url (it will only work for http though).

Fixes bug #461652

svn path=/trunk/; revision=7517