Fix missing HTTP information when opening links in new tabs
Bug #120341

This changeset reworks ephy_shell_new_tab_full (and its callers) to
use a WebKitNetworkRequest instead of an URL when handling openning
links in new tabs. This gains us the advantage of feeding the new
WebView with the same WebKitNetworkRequest WebKit would use to
navigate to the new address.