Performance: Set a fixed width on autosized cell renderer
The text cell renderer in use by the url autocompletion is the only
renderer that expands. As the width of the entry completion is known in
advance (as big as the entry you are completing on), we can set a fixed
width. Since the height of the cell renderer is hardcoded, too, the
complete size of the cell renderer is known in advance and Gtk skips
layouting the text to compute the size.

This is a noticable performance improvement, expecially when the
completion contains lots of visible entries.