Move address normalization to the EphyWebView functions.
EphyLink is called repeatedly when opening URLs, so normalizing the
address there all the time is pointless. Also, we don't go through
here in all cases, so we need to add the normalization in the
ephy_web_view_load_request function too; this fixes opening URLs like
"foo.com" from the command line, for example.