ephy-embed.c: workaround possibly bogus COMMITTED load-status from WebKit
When WebKitGTK+ fails to load a URI and is allowed to load the default
error page it will emit a COMMITTED load-status with the original URI
we failed to load. This was confusing epiphany, since in theory
COMMITTED should be only emitted with URIs known to be good. As a
workaround flag pages that fail to load in the callback for
'load-error' and avoid doing anything with them when they arrive in
COMMITTED.

Bug #593200