ephy-location-action.c: unblock address sync on switch-page too
We need to unblock the address syncing for the location on switch-page
too, since it's not guaranteed that we'll receive a focus-out or
activate event before we are interested in resyncing the URL entry
again. For example, if the focus is on the entry and we switch tabs
with the keyboard shortcut the entry will go out of sync, keeping the
URL of the previous tab.

Bug #594346