Bring back epiphany context menu
There's a few items (like email link) and actions (like bookmark link)
missing or not working because of missing information in the
WebKitHitTestResult object, but most of the stuff is working.

For some reason the g-ir-scanner is not picking up the correct type
name for WebKitHitTestResult (it uses WebKitHitTestResult instead of
WebKit.HitTestResult), so the introspection support is broken unless
that error is fixed manually. Looking into that ...

Bug #562617