Don't use deprecated GtkButton API
In 82a5da33 we removed usage of gtk_button_pressed and gtk_button_released
however replacing those for g_signal_emit calls of button-press-event and
button-release-event. This however caused a crash on middle clicking of
elements in the toolbar.
To fix this we use g_signal_emit with pressed and released signals, these
two are deprecated signals, though.

There doesn't seem to be a way to trigger button pressed/released animation
without the pressed/released *deprecated* signals or functions, so this is our
best solution for now.

Bug #603450