Disable introspection for distcheck
There's a (apparently) nasty circular dependency issue when doing
distcheck with introspection enabled. The .gir file needs the epiphany
binary to be built, so basically when doing the dist part of check
copying the files to the temp directory we trigger a full build. This
would be already bad, but on top of that it does not work, since we
use BUILT_SOURCES for some generated files, which does not work during
'make dist'.

So disable it while we figure this out.