Default to one toolbar with less buttons
This change follows from the GNOME decision to default to "Text
besides icons", which makes the first toolbar smaller
horizontally. With the woohoo bar the need for history and bookmark
buttons is also diminished, given that you can easily search them
using the location bar itself.

Zoom buttons are kept, but moved to the other side of the toolbar,
after the location bar, to clearly distinguish them from navigation
operations.

Bug #412385