downloader-view: stay alive until notifications are gone
When using notifications for downloads we usually hit the ugly case where the
notification pops but there is no GtkStatusIcon for it so it will pop in the
default location (usually bottom right). This is inconsistent with the behavior
when the GtkStatusIcon is present, which is to show it attached to it.

To fix this we hold a reference to the DownloaderView when showing the
notification and release it when such notification has been closed.

Bug #611779