downloader-view: use update_download_row correctly on add
When adding a download we should call update_download_row() *after* connecting
the signal handlers or they will never be disconnected.
Also, in update_download_row() we should consider that the case where it is
called with a download that has already been removed as a non fatal case (i.e.
just return;). This is because in downloader_view_add_download() we can have
the signal handlers remove the download from the view before the manual call
(it was happening vice versa before this).