migration: be more strict with checks when storing
Make sure we don't confuse form passwords with HTTP auth passwords. And also
check that username is not empty (at least that). Sometimes decryption fails
(not even nss tools work) and we get empty usernames. As you can see, we are
assuming here that usernames are always mandatory but passwords might not be.