Split EphyNavigationAction in one abstract class and two subclasses
To ease understanding and further modification of the two different
usages for the EphyNavigationAction class (Back/Forward and Up buttons),
all the code there was split so the common one is kept in the
superclass, delegating the more specific parts in the implementations
of the subclasses: History (back/forward) and Up buttons.

Also updated usage of this class in EphyToolbar.

Bug #539716

Signed-off-by: Xan Lopez <xan@gnome.org>