ephy-location-entry: reenable favicons in completion
Favicons in the completion popup were hidden because of an ugly flashing. It
turns out it is caused by gtk_entry_set_icon_from_* functions. They queue a
resize if the GtkEntry is visible, hence we see this ugly flash which is
actually the completion obeying to the resize.

The side effect is that when you start typying a new address in the location
entry, the favicon of the current location will still be displayed. The
previous behaviour was to set a 'text/html' icon when the user edited the
location entry. Note that such icon will disappear as soon as you activate the
location entry.

Bug #616345