Use the Geolocation support in WebKitGTK+
Using the infrastructure for info bars, present policy requests to the
user so they can allow/deny them based on the URI of the frame that is
requesting. Moving forward, it would be a good idea to offer to save
the decision, and store it in the history entry for that site.

Bug #626687