ephy-download: add the new EphyDownload object
EphyDownload is a wrapper object around WebKitDownload that handles common
behavior in downloads: auto-destination, default action for the MIME type.

It can be used to wrap a WebKitDownload coming from a WebKitView or to download
a url: ephy_download_new_for_uri and ephy_download_new_for_download are
provided.

Its lifetime is not automagic like EphyEmbedPersist, so you have to unref it
when you no longer need it.

This new object replaces EphyEmbedPersist and enables us to use a single
codepath for downloads in all Epiphany.

Bug #618443