Better behaviour when openning new windows that are resized as tabs
When a new window is opened and given a specific size and position by
javascript, the request should only be respected if the new window is
being opened stand alone, and considered a popup. This fixes bad
behaviour observed when the new window became a tab on an existing
window and messed with its size.

Bug #612155