Mark windows with modified chrome as popups and only allow resizing for them
A window with non-modified chrome will potentially be used as a normal
window for browsing - there's no reason why it should be different
from existing windows when it comes to resize policy. With this change
we consider windows with modified chrome as popups and only set a
default size on creation for popup windows.

Bug #612155