toolbareditor: Fix the removal of items by DnD.
If the 'window-dragging' property of GtkToolbar is set (as it is in adwaita)
then the Toolbar starts the drag of the window on 'button-press' event.
We prevent this by connecting a handler for 'button-press' which prevents
the handler of the toolbar to be run while in edit mode.

Fixes bug #647265