Merge EphyApplication into Ephy(Embed)Shell
We want to use EphyApplication from embed/ too, so instead of
following the EphyEmbedShell/EphyShell pattern just fold the code into
the Shell itself. This makes sense since both classes have basically
the same function, and we can easily leave the non-embed code in
EphyShell and move to EphyEmbedShell the bits we want to be global.

The better name would be EphyCoreApplication/EphyApplication for the
resulting object, but we can do this later.