ephy-window: clear finished downloads when closing the downloads bar
When a download finishes but you don't 'activate' it (open or browsing) the
widget stays alive on the downloads-bar, worse, if open or browse actions fail,
the download can't be removed from the bar.

This commit clears finished downloads from the downloads-bar when it is closed
(with the X button). Hiding it will *not* trigger this cleanup.

Bug #653918