e-file-chooser: remove persist-key
GTK+ has changed the UX of the GtkFileChooser. This has (pragmatically)
deprecated "last save dir" stored by applications themselves.

For Epiphany this means we no longer need a "persist-key" in our file
chooser, and that we have to clean up some GtkFileChooser API use in
followup patches.

This commit removes the property and uses of it.

Bug #655508