ephy-shell: do not resume session in web app mode
Session resume will ask to resume the session (opened pages, etc) if
the previous instance crashed. We don't really want this in web app
mode, so skip it. Note that this is not exactly the same than loading
the session, which just restores the previous window sizes and
states. We do want that, but the terminology here could be more clear.

https://bugzilla.gnome.org/show_bug.cgi?id=658739