e-download-widget: update icon every 10%
In some cases trying to get the content-type of a download right at the
start of the transfer won't give us useful information.

To handle this case we query for the content-type and a proper icon
every 10%, hoping that GIO/Soup will make a smarter guess. This is an
arbitrary number that hopes to work with most cases.

Bug #662059