web view: fix icon names
The GTK+ 3.0 themes have a different name for the dialog icons
(dialog-{information,error}) rather than the stock GTK+ icon names
(gtk-dialog-{info,error}).  Use the proper names.

Additionally, ensure that icon_info is not NULL when performing the lookup,
which will fail.  This part of the change is to avoid a possible crash/segfault
caused by the lookup without a valid icon_info.

Bug: 665170
Signed-Off-By: Saleem Abdulrasool <compnerd@compnerd.org>