Remove custom EphyToolbar in favor of a vanilla GtkToolbar
This allows us to stop using EggEditableToolbar here, we just load a
normal toolbar from GtkUIManager in EphyWindow. The special toolbar
actions and methods have been moved to EphyWindow too, mostly in a
mechanical fashion (probably some simplifications are possible).

The code in the bookmarks UI that depended on EphyToolbar has been #if
zeroed instead of porting it, since it will go away in the near future
too.

https://bugzilla.gnome.org/show_bug.cgi?id=664483