history-service: make sure we can update pages' titles and urls
When we retrieve a url from the backend from a user given
EphyHistoryURL, it's possible for this to contain updated information
for the page. In such case, don't overwrite this with the one stored
in the backend.

This makes possible to later update this information when saving a new
visit.