ephy-history-service: add 'visit-url' signal
Similar to the old EphyHistory 'add-page' signal. It allows blocking
the visit by handling the signal, since the actual logic to add the
visit is in the class' default handler.

The add_visit(s) methods won't emit this signal, but for now we'll
assume that whoever is using those knows what he's doing.