ephy-history-service: make the async API cancellable
The main purpose of the cancellable API in ephy-history-service
is to let the user notify when the results of the operation
are no longer needed and the callback call can be omitted. Since
performing a read operation, in such cases, makes no sense, we
cancel them altogether. However, given the nature of the service,
we don't cancel write operations.

https://bugzilla.gnome.org/show_bug.cgi?id=671901