Fix distcheck
Most data files are listed through EXTRA_DIST, so they are left in the
top srcdir during VPATH builds. The new EphyWebView unit test needs
data/pages/error.html, so distcheck was broken because ephy_file
("error.html") was failing since it looked for them in $pwd/data.

To fix this, look for files in the top srcdir instead of $pwd/data,
which will lead to the builddir data data dir during distcheck.