e-location-entry: make aligment pixel-perfect
Align the elements of the GtkEntryCompletion popup with those in the
location entry. The code comes with a detailed comment and a scheme of
how the aligment is done now.

Because of the unhandled pixels of GtkEntryCompletion, this code might
need an update if anything in GTK+ or Adwaita changes.

https://bugzilla.gnome.org/show_bug.cgi?id=672927