Add a setting to control whether the session is automatically restored
We add a new gsettings key, 'restore-session-policy', with two valid
values: 'always' and 'never'. A brief explanation of our session state
mechanism follows.

There are three ways to exit Epiphany:

1) Activate 'Quit' in the application menu
2) Close the last application window
3) Kill the process manually, SIGSEGV, or other similar unexpected
event.

For 1) and 2), we'll now do the same thing:

a) Call ephy_session_close
b) Exit the application manually

ephy_session_close will check the new restore-session-policy setting,
and only save the session state if it's set to 'always'. Before it
used to manually destroy all present windows. We now let EphyShell or
EphyWindow do this, EphySession only manages the session state saving.

For 3), the process will die with the state saved up to that point,
there's nothing we can do. For that reason, on startup also check the
new setting; if it's set to 'never' ignore the session state, open a
window in the homepage, and delete the old state file.

https://bugzilla.gnome.org/show_bug.cgi?id=673453