Allow for more fine-grained file helpers init
Since we are about to migrate our profile dir, allow file helpers init
to not ensure the profile dir exists (it was hardcoded until now). For
this we get rid of the ugly boolean parameters and add a flags
parameter, which preserves the old behaviors and allows for this new
option.

We update all the callers in the tree.