ephy-profile-migrator: migrate Web Applications to new profile location
We need to update their desktop files (they had references to the
profile directory) and the symlink of the .desktop file in the Shell
applications directory.

It would be much easier to just delete and re-add the applications,
but unfortunately that would wipe out the existing profile data in the
apps (like cookies).