tests: add the beginnings of ephy-session unit testing
For now a simple test that verifies that loading a session works.

We need to link with the resource files, which is unfortunate, but I'm
not sure there's a way around it since this test will actually spawn a
full EphyWindow (ie, it's not only testing the session or shell code
but the entire browser as a side effect).