ephy-web-view: try to simplify a bit the typed-address logic
I think we can do without the "expire_address_now" boolean, and just
rely on whether typed-address itself is NULL or not. If it is not, we
want to always show it in the location entry and never overwrite it,
since it reflects what the user has typed. The only case when we want
to go ahead and set it to NULL is when the load actually starts (ie,
when the user wants to transform what he has typed into a new
load). Try to document things better reflecting this logic.