Move windows handling code from EphySession to EphyShell
EphyShell is now a GtkApplication that already does window handling,
however we want to maintain only the of browser windows. Instead of
keeping the window lists in ephy-session, it's easier to override
window_added and window_removed virtual methods of GtkApplication and
update the browser window lists. Most of the cases where this API was
used, had to get the session object from the shell first, now we save
this step keeping the code simpler.

https://bugzilla.gnome.org/show_bug.cgi?id=641734