Use webkit_web_frame_load_alternate_string to load error pages
webkit_web_view_load_string, which was used previously, does not add
the error page to the back forward list. This means that when you
go back from an error page (n) you will skip directly to an older page
(n-2) or not be able to go back if the error page was the first loaded.