ephy-window: centralize the logic about invisible URIs in one place
Let's make EphyWindow the one in charge of deciding whether a URI is
actually shown or not in the location entry. This allows to remove
some code to that effect in EphyLocationController (and perhaps some
more in EphyWebView in the future), and makes this feature more
extensible for the future.