ephy-web-view: avoid getting a 'Blank page' title for non titled pages
In WebKit2 we'll get no notify::title for pages without title, so
check this case and reset the title to NULL, which for non-blank pages
will use the address as title.

This also removes the last notify::embed-title calls outside of
ephy_web_view_set_title.

Cherry-picked from a patch by Carlos Garcia.

https://bugzilla.gnome.org/show_bug.cgi?id=682354