ephy-overview-store: use the age of a thumbnail only to decide when update is needed
Currently, we were not retrieving thumbnails from the service if they
were outdated. This would cause some pages in the overview not to
display a thumbnail at all even if one is available (but old).

Fix this by always retrieving a thumbnail but storing its mtime in the
model, and making ephy_snapshot_store_needs_snapshot() check the age
of snapshots when deciding whether a snapshot update is needed or not.