ephy-profile-migrator: allow running migration in non-default profile directories
This is necessary since we add new columns to the history database in
step 8 of migration. However, non-default profile directories don't
have a .migrated file right now but, since up to step 7 it was
possible to use these profiles without migration, it is safe to assume
that a non-default profile without a .migrated file can be migrated
starting from step 8.

https://bugzilla.gnome.org/show_bug.cgi?id=681679