ephy-snapshot-service: Split ephy_snapshot_service_get_snapshot_async()
ephy_snapshot_service_get_snapshot_async() receives an option web view
parameter, that it's only used in case the snapshot is not the in the
thumbnails cache. We can split the method into
ephy_snapshot_service_get_snapshot_async() to get a snapshot from a web
view and ephy_snapshot_service_get_snapshot_for_url_async() to get a
snapshot from the cache. The former uses the latter to try first if the
web view URI is in the cache.

Patch includes other cleanups and fixes:

  - Add missing ephy_snapshot_service_save_snapshot_finish()
  - Add EphySnapshotServiceError to handle errors
  - Use GSimpleAsyncResult API instead of using GIOScheduler directly
  - Use different async data structs for every async operation

https://bugzilla.gnome.org/show_bug.cgi?id=683327