Use the new EphyHistoryService::urls-visited signal
Since both the frecent store and the history window don't need to be
updated immediately after a url is visited, we can use ::urls-visited
instead. The advantage of this is that we reduce considerably the load
when updating both models when loading many pages at the time,
specially during startup.

https://bugzilla.gnome.org/show_bug.cgi?id=683550