ephy-session: respect 'new-windows-in-tabs' setting when opening links
Right now invoking 'epiphany <URI>' when there's already a browser
instance running would always open a new window, unless a new tab was
explictly requested with -n/--new-tab. We should honor the
'new-windows-in-tabs' setting (which is TRUE by default), and open the
link in a new tab in the last active window if no special conditions
were specified.

https://bugzilla.gnome.org/show_bug.cgi?id=641949