Focus windows when opening new links in tabs
If someone requests a link to be opened by ephy in an already running
instance, and that link is opened in a new tab (either because of our
settings or because -n/--new-tab was passed), the window where this
happened will not be presented, so it might go unnoticed.

Fix this by presenting the windows where new tabs are opened through
the open_uris mechanism in EphySession.

https://bugzilla.gnome.org/show_bug.cgi?id=641949