Import adblock extension into the source tree
This commit imports the adblock extension from 'epiphany-extensions'
with as few modifications as possible. Further modifications and
simplifications are possible, but for now we just load it
unconditionally, enabled.

We have it in the src/ directory because it depends on some high-level
objects in Epiphany (EphyExtension, EphyWindow), but in the future
we'll probably move it to embed/, transforming it from an extension
into just the specific implementation of our EphyAdblockManager
object.

https://bugzilla.gnome.org/show_bug.cgi?id=681657