Fix a ton of 'warning: function declaration isn’t a prototype'
New warning flags have been added to gnome-common recently, and we
were getting this a lot. Turns out in C 'foo ()' is not the same than
'foo (void)'; the first just means that no information is given about
the number of arguments, the second means the function has exactly
zero arguments, so add the 'void' thing all over the place when
needed.