Append new tab on the window with most tabs in the current workspace
This change uses a different heuristic to decide on what the best window to add
a new tab is from the last window which has been interacted with, which might
even be in a different workspace, to the window with the most tabs in the
current workspace. If no window exists on the current workspace one will be
created.

Partial fix for https://bugzilla.gnome.org/show_bug.cgi?id=685976