ephy-session: fixup session autoresume
This code was confusing, but what we actually wanted to do here was to
only autoresume the session if there were no windows opened, otherwise
it means we are launching a new instance after the first one.

Make this explicit in the code instead of relying on side-effects of
early returns.