ephy-state: rename to 'ephy-initial-state'
The purpose of EphyState is to track the sizes and positions of
windows, paned or expanders in order to remember what is the preferred
*initial* and *default* state of those UI elements. So for example we
merge the tracking of the size/positions of all main windows in one
record, because we only need an initial value which we'll use as
default for newly created EphyWindows.

Since this is a very specific task, different to the actual tracking
of all sizes and positions in EphySession in order to restore them at
startup, rename the class to EphyInitialState to avoid confusions.