ephy-download: Make EphyDownload:window a construct only property
It's always set right after the download object is created and it
doesn't have any effect when it's set after the download has been added
to the embed shell. This way we can make sure that when the download is
added to the embed shell the parent window has already been set if
there's any.