Only load pages when their tab is switched to upon session restore
Firefox has led the way implementing this behaviour to improve the experience
of restoring a session with lots of tabs. By delaying the loading of pages to
when the user shows interest in them, the time it takes for the browser to
become usable is diminished, and less pages are loaded in parallel, which
improves load time for the first pages the user sees.

It also has the advante of displaying less HTTP Basic Auth dialogs, when the
user has many tabs pointed to the same server.

https://bugzilla.gnome.org/show_bug.cgi?id=675302