Fix loading HTTP auth protected pages from the overview
The switch between overview and normal web page happens when the page
load is committed, in case of HTTP auth protected pages we never switch
to web page mode because authentication happens before the load is
committed. In WebKit2 the auth dialog is attached to the web view, so we
need to show the web view before the load is committed to show the auth
dialog.

https://bugzilla.gnome.org/show_bug.cgi?id=694268