Add setting for delayed tab loading, and use it to fix the session tests
The session tests were broken by the delayed tab loading feature - the restored
embeds would not finish loading and thus the expectations for the URIs were not
met. This change adds a setting that is now used by those tests to disable the
feature temporarily.

Also revert "ephy-session-test: make tests pass", which is no longer necessary.

This reverts commit 75da5fff3f4489dacf4ded6f012daa06af306709.

https://bugzilla.gnome.org/show_bug.cgi?id=694470