ephy-session-test: make session tests work more reliably
Based on a patch by Manuel Rego.

Ensure the WebViews created during the session load are fully loaded
before we check their URIs. This was working more or less by pure
chance in WK1, and was failing in WK2 because of the process
separation.

The restore tab test is skipped in WK2 for now, since it's crashing.

The patch excludes some asserts in WK2 while bug #695437 is not fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=695300