ephy-profile-migrator: Stop spinning when porting form passwords
Use atomic operations to manipulate the counter that tracks the number
of operations we need to achieve, to ensure coherency when the results
are received from different threads.

Also use our own main loop rather than spinning on g_main_context_iteration()
to avoid CPU burning when waiting for completion.

Finally, fix the migrator never returning as the number of operations
was not decreased on the ephy_form_auth_data_store() error path.

https://bugzilla.gnome.org/show_bug.cgi?id=695828