Load back item when back is clicked with control held/middle clicked, for wk2
In WebKit2 the back/forward list is not copiable, so that assumption
doesn't hold, and the current code path fails. With this change we
tell EphyShell to not copy the list and load the URL held by the back
item of the existing WebView in the new WebView instead, for WebKit2.

https://bugzilla.gnome.org/show_bug.cgi?id=695951