Move code to initialize preferences from ephy-main to ephy-shell
As well as the code the prepare the environment for the wen process and
web context initialization. This allows to only run that initializations
only for the primary instance. It fixes a crash in web process when the
ui process finishes too early and makes secondary instances to run
faster, since they don't spawn a web process anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=696020