ephy-web-view: disable modified forms check in WebKit2
This code is causing some problems because the check happens in a sync
call right before a WebView is destroyed. There seems to be a race
condition and some times the method call will hang for seconds waiting
for a dead process on the other side.

We'll figure it out for 3.8.1, but for now disable it since it's
causing more harm than good.