prefs: Use the correct signal to monitor GtkFileChooserButton's selection changes
The correct signal is 'selection-changed', as in all file choosers.

The signal that was being used, 'current-folder-changed', only means that the current
directory being shown in the file chooser dialog has changed - not that the selection
has changed.

See https://bugzilla.gnome.org/show_bug.cgi?id=696498#c8 - although that comment
suggests the use of the 'file-set' signal, which I don't think is very useful.  It's
just a historical remnant.