ephy-string: Avoid poking at remote websites
Epiphany tries to query the existence of remote sites before loading
their page. This causes problems when the remote site is not available
and loading it requires the query to timeout (epiphany would look like
it's not loading the site for up to 60 seconds) or for one-time URLs
(where it might invalidate the query, such as for password resets, or
confirmation e-mails).

We'll now only check whether the file exists so that incomplete URLs
("epiphany foo.com/bar.html") still loads the bar.html file if it exists
locally, but the remote website if not.

https://bugzilla.gnome.org/show_bug.cgi?id=703772