Add a new startup flag to indicate we are resuming a session
Web should open a new window if no URIs are passed - it's how GNOME
Shell does its 'new window' magic. However, it should not do so if the
instance that is already running is a headless instance that happens
to be active, and resumes a session when called with or without URIs.

https://bugzilla.gnome.org/show_bug.cgi?id=707451