Change default charset to iso-8859-1 for greater web compatibility
When servers reply with no explicit charset and the page does not specify a
charset either, the default behaviour on the web has been to consider the
contents to be in the iso-8859-1 encoding. That is why WebKit's default charset
setting defaults to iso-8859-1. By changing that default to UTF-8 we break
some web sites, such as http://www.planalto.gov.br/ccivil_03/leis/lcp/lcp101.htm.

This change restores web compatibility by using the same default as WebKit.

https://bugzilla.gnome.org/show_bug.cgi?id=720247