Mark keyword-search-url default for translation
We allow region parameters in the default "Search the web" smart
bookmark, which probably nobody ever uses, so why not in the default
search engine that people actually DO use? Also needed because the DDG
URL in the preferences dialog will be translatable, and this must match
it to avoid being detected as a separate, custom search engine.

https://bugzilla.gnome.org/show_bug.cgi?id=709459