Use a flag to decide if a tab is download-only
There are some cases in which a view may not be able to go back and still be not
just for a download: if the view has been opened by middle clicking somewhere or
by having an URL provided on startup, then the navigation from the overview to
the first page won't have happened. By tracking whether the page has ever
committed a load we can be sure it has been opened just for the download.

https://bugzilla.gnome.org/show_bug.cgi?id=720609