downloads: Do not add automatically the downloads to the UI when they are created
This allows the creator to decide whether to add the download to the UI
or not and when. The downloads are no longer handled by the
EphyEmbedShell making clear who takes the ownership. Downloads are now
explicitly added to a particular Ephywindow, instead of all the windows
be notified by a download and deciding whether such download belongs to
that particular window or not.