ephy-find-toolbar: properly set its webview
We were setting the webview twice to the find-bar: upon creation and
then everytime the notebook switched to the page containing it. However,
only the second time all the signal handlers related to search
were connected, and not even properly.

Fix all this mess by using the proper setter() and removing the public
method, after all the webview property is construct only.

https://bugzilla.gnome.org/show_bug.cgi?id=724230