Decouple the page load from the tab creation
The flags used to decide whether to load the page and how are confusing,
because they are mutually exclusive, which adds more complexity to
ephy_shell_new_tab. Also, ephy_shell_new_tab doesn't always create a new
tab, depending on the parameters and flags, it might reuse an existing
empty tab. All this is easier to handle from the callers if we make
ephy_shell_new_tab always create a new tab that is returned and callers
can decide what to load in the new tab.