Replace the location bar with the title of the web page
From the moment there is a title for the web page, hide
the location bar. Instead, show the title of the web
page as title of the header bar and the url of the web
page as subtitle.

Clicking on the title (when it is shown) will switch to
the location bar.

Type Esc when the entry of the location bar is focused,
or focus on the web view will switch to the title.

Reviewed by Carlos Garcia Campos <carlosgc@gnome.org>
and William Jon McCann <william.jon.mccann@gmail.com>.

https://bugzilla.gnome.org/show_bug.cgi?id=711408