Add ephy-resource:// custom URI scheme implementation
Ideally we could use resource:// URIs since they are supported by both
WebKit and libsoup, but the web process doesn't have access to the
resources compiled in the UI process. And in multiprocess mode, the
network process doesn't have access to the resources either.
ephy-resource:// URIs are like a proxy to load gresources compiled in
the UI process, the data is sent directly to the web process in
single process mode or to the network process in multiprocess mode.

https://bugzilla.gnome.org/show_bug.cgi?id=724862