embed-utils: Add ephy_embed_utils_address_is_valid
Helper method used to check if a url is valid to decide whether to load
it or perform a search. It's mostly the same expression used by
EphyWebView, but moved to a helper function. The regular expressions used
in that expression are also moved to embed-utils and they are only
compiled once instead of everytime a web view is created.

https://bugzilla.gnome.org/show_bug.cgi?id=725081