Update help index and introduction
Index <desc> is now used as the project description on h.g.o/users, so
should contain something of interest to users and "epiphany" because
"web" is a poor search term on the internet.

Using epiphany will not make the web any more nicer a place. Would be
good to mention something about privacy as it's a moderately big
feature. Can't think of a nice way to fit it in right now.

<p> inside <media> are alt text, used to describe the image and double
up as instructions for retaking the screenshot.