Fix the build with new WebKit
New WebKit aborts the compilation if webkit2.h and
webkit-web-extension.h headers are included together. We are including
the ephy-web-extension.h header from the UI process only to get the
names of the dbus service, interface and path, so move those defines to
its own file that can be included by both UI and web process files.