ephy-web-view: do not clear typed address on first load
Previously, the first page loaded in a webview would be done with
before the user had any chance to type an address, so it was safe to
assume that any typed address could be cleared during page loads,
as they would always come from previously loaded pages.

Now that the process of creating a webview is different and
involves spawning a web process that needs to be ready before
the initial page can be loaded, it might be possible for users
to type an address before this is ready. This breaks the previous
assumption and needs to be dealt with, in order to avoid
cleaning up typed text during the initial page load.

Additionally, it makes sense to make set_address() a no-op if
the given address is the same as the currently loaded.

https://bugzilla.gnome.org/show_bug.cgi?id=728143