ephy-uri-helpers: Dot no modify the URI when it doesn't have garbage
We are currently decoding/encoding the uri query and returning it
unchanged when it doesn't have garbage. We should return NULL instead in
such case.
This fixes the session management issues in some sites like gmail, but
because the URLs processed don't actually have garbage. A URL with
garbage and an underscore will still fail. This patch adds a unit test
cover that particular case.

https://bugzilla.gnome.org/show_bug.cgi?id=730464