Set the history table sortable by any of the columns. https://bugzilla.gnome.org/show_bug.cgi?id=699519
After the history service got support for ordering history results
by various columns, update the history dialog to allow sorting.
* override the column header clicked event to change the sorting
order and/or direction
* reload the data from the history service on column header clicks

In case the history window contents are resorted, to avoid freezing
the user interface, do the following:
* register a source to process the url list received from the service
multiple urls at a time, (as many as possible before the next frame
 drawing) by adding them to the model, and freeing the url
* in case the full list is loaded or another sort is requested, remove
the source, and free the remaining list
* as clearing the liststore while it is set to a treeview emits the
row deleted signal for each row, the liststore is set to null before
clearing, it is cleared, and then set again for the treeview
* as setting the model again resets the sort column and indicator
on the treeview, those attributes must be set again.